% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{autophagySign}
\alias{autophagySign}
\title{Autophagy Signature}
\usage{
autophagySign(
  dataset,
  nametype = "SYMBOL",
  author = "Xu",
  hgReference = "hg38"
)
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows
correspond to genes and columns correspond to samples.
Alternatively, an object of type \linkS4class{SummarizedExperiment},
\code{\link[SingleCellExperiment]{SingleCellExperiment}},
\code{\link[SpatialExperiment]{SpatialExperiment}} or
\code{\link[SeuratObject]{SeuratObject}}.}

\item{nametype}{character string saying the type of gene name ID (row names
in dataset). Either one of "SYMBOL", "ENTREZID" or "ENSEMBL".}

\item{author}{character string saying the first author of the signature
publication. Check it in \code{\link[signifinder]{availableSignatures}}.}

\item{hgReference}{Human reference genome: "hg19" or "hg38"}
}
\value{
A SummarizedExperiment object in which the Autophagy score
is added in the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function
 \code{\link[signifinder]{availableSignatures}}.
}
\examples{
autophagySign(dataset = ovse)

}
