% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{availableSignatures}
\alias{availableSignatures}
\title{Show available signatures}
\usage{
availableSignatures(
  tumorTissue = NULL,
  signatureType = NULL,
  datasetInput = NULL,
  description = TRUE
)
}
\arguments{
\item{tumorTissue}{filter for tissue type}

\item{signatureType}{filter for signature type}

\item{datasetInput}{filter for input type}

\item{description}{whether to show or not the signature description}
}
\value{
a data frame
}
\description{
It shows a table containing all the information of the signatures collected
in the package.
}
