% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{cellCycleSign}
\alias{cellCycleSign}
\title{Cell-cycle Signature classifier}
\usage{
cellCycleSign(
  dataset,
  nametype = "SYMBOL",
  author = "Lundberg",
  inputType = "microarray"
)
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows
correspond to genes and columns correspond to samples.
Alternatively, an object of type \linkS4class{SummarizedExperiment},
\code{\link[SingleCellExperiment]{SingleCellExperiment}},
\code{\link[SpatialExperiment]{SpatialExperiment}} or
\code{\link[SeuratObject]{SeuratObject}}.}

\item{nametype}{character string saying the type of gene name ID (row names
in dataset). Either one of "SYMBOL", "ENTREZID" or "ENSEMBL".}

\item{author}{character string saying the first author of the signature
publication. Check it in \code{\link[signifinder]{availableSignatures}}.}

\item{inputType}{character string saying the type of data you are using.
Either one of "microarray" or "rnaseq".}
}
\value{
A SummarizedExperiment object in which scores are added in
the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function
 \code{\link[signifinder]{availableSignatures}}.
}
\examples{
cellCycleSign(dataset = ovse, inputType = "rnaseq")

}
