% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{CINSign}
\alias{CINSign}
\title{Chromosomal instability Signature}
\usage{
CINSign(dataset, nametype = "SYMBOL")
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows
correspond to genes and columns correspond to samples.
Alternatively an object of type \linkS4class{SummarizedExperiment},
\code{\link[SingleCellExperiment]{SingleCellExperiment}},
\code{\link[SpatialExperiment]{SpatialExperiment}} or
\code{\link[SeuratObject]{SeuratObject}} containing an assay
where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset (row names).}

\item{typeofCIN}{A pre-defined number 25 or 70. It represents the number of
genes on which the score can be calculated. The 25 genes represent a subgroup
derived from the 70.}
}
\value{
A SummarizedExperiment object in which the score will be
added in the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function
 \code{\link[signifinder]{availableSignatures}}.
}
