% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{CorrelationPlot}
\alias{CorrelationPlot}
\title{Correlation Plot}
\usage{
CorrelationPlot(data)
}
\arguments{
\item{data}{dataframe con le signature sulle colonne oppure oggetto con i colData}
}
\value{
A correlation ellipse graph
}
\description{
Correlation Plot
}
