% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{.barkleyFun}
\alias{.barkleyFun}
\title{Pancancer Cellular States Barkley Scoring Function}
\usage{
.barkleyFun(dataset, signList, modules)
}
\description{
Pancancer Cellular States Barkley Scoring Function
}
