% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{EMTSign}
\alias{EMTSign}
\title{Endothelial-Mesenchymal Transition Signature}
\usage{
EMTSign(dataset, nametype, ...)
}
\arguments{
\item{dataset}{A matrix of expression values where rows correspond to genes and columns correspond to samples. Row names must be Official Symbol.}

\item{...}{Other arguments passed on to the GSVA function.}
}
\description{
Given a dataset, it returns the Endothelial score and the Mesenchymal score for each sample, based on QH Miow (2015).
}
