% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{GeneExpressionHeatmap}
\alias{GeneExpressionHeatmap}
\title{Scatterplot for a single signature}
\usage{
GeneExpressionHeatmap(data, signatureName, dataset)
}
\arguments{
\item{data}{either a data frame or a SummarizedExperiment.}

\item{signature}{If data is NULL, signature must be a numeric vector. When data is provided,
signature must be the character string of the column name in data.}
}
\value{
A ComplexHeatmap object
}
\description{
Given a signatures...
}
