% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{geneHeatmapSignPlot}
\alias{geneHeatmapSignPlot}
\title{Scatterplot for a single signature}
\usage{
geneHeatmapSignPlot(
  data,
  whichSign,
  logCount = FALSE,
  splitBySign = FALSE,
  sampleAnnot = NULL,
  splitByAnnot = FALSE,
  ...
)
}
\arguments{
\item{data}{output from a signature function}

\item{whichSign}{indicare quali signature usare, di cui verranno presi i geni}

\item{logCount}{whether or not to compute logarithm for the expression values in the dataset}

\item{splitBySign}{se splittare o no le righe dell'heatmap in base alla signature di provenienza dei geni}

\item{sampleAnnot}{un vettore di annotazione dei sample}

\item{splitByAnnot}{se splittare o no le colonne in base all'annotazione dei sample, in
questo caso sampleAnnot deve essere una variabile categorica}

\item{...}{Heatmap params}
}
\value{
A ComplexHeatmap object
}
\description{
Given a signatures...
}
