% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{GetAggregatedSpot}
\alias{GetAggregatedSpot}
\title{7 spots}
\usage{
GetAggregatedSpot(dataset)
}
\arguments{
\item{dataset}{Seurat object of a 10X Visium dataset}
}
\description{
Given a 10X Visium dataset, it reassigns to each spot the aggregation of it with the nearest.
}
