% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{getSignGenes}
\alias{getSignGenes}
\title{Get Signature Gene List}
\usage{
getSignGenes(whichSign)
}
\arguments{
\item{whichSign}{name of the signature. The names are those in column
'signature' from the table which is obtained by
\code{\link[signifinder]{availableSignatures}}.}
}
\value{
A dataframe object with "SYMBOL" in the first column. Some signatures
have also additional colums: "coeff" for coefficients that weigh the gene
contributions; "class" for a classification that divides the signature in
two or more groups. Few signatures have other specific columns.
}
\description{
This function returns the list of genes of a signature.
}
\examples{
getSignGenes("EMT_Miow")

}
