% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{heatmapSignPlot}
\alias{heatmapSignPlot}
\title{Global Heatmap of Signatures' scores.}
\usage{
heatmapSignPlot(
  data,
  whichSign = NULL,
  clusterBySign = NULL,
  sampleAnnot = NULL,
  signAnnot = NULL,
  splitBySampleAnnot = FALSE,
  ...
)
}
\arguments{
\item{data}{An object of type \linkS4class{SummarizedExperiment} coming from
signatures functions. Th function takes as input a matrix where each
row represents a signature and each column represent a samples.}

\item{whichSign}{The signatures to show in the plot.}

\item{clusterBySign}{One ore more signatures that clusterize columns on
Heatmap.}

\item{sampleAnnot}{A vector containing samples annotations.}

\item{signAnnot}{A character vector of signature annotations: signature,
topic, tumor, tissue.}

\item{splitBySampleAnnot}{A categorical variable. Give you the possibility
to split columns based on samples' annotations.}

\item{...}{Other parameters specific of the function
\code{\link[ComplexHeatmap]{Heatmap}}.}
}
\value{
A \code{\link[ComplexHeatmap]{Heatmap-class}} object.
}
\description{
Given a matrix of multiple signatures, the function gives the opportunity to
observe the trend of the signature's scores for each samples. Moreover it
gives the possibility to order all the matrix based on a single signature.
}
