% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{HypoSign}
\alias{HypoSign}
\title{Hypoxia Signature}
\usage{
HypoSign(dataset, nametype)
}
\arguments{
\item{dataset}{expression values where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset row names.}
}
\description{
Given a dataset, it returns the hypoxia score for each sample as in Buffa et al. 2010.
}
