% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{IPRESSign}
\alias{IPRESSign}
\title{IPRES Signature}
\usage{
IPRESSign(
  dataset,
  nametype = "SYMBOL",
  tumorTissue = "skin",
  pvalues = FALSE,
  nperm = 100,
  ...
)
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows
correspond to genes and columns correspond to samples.
Alternatively an object of type \linkS4class{SummarizedExperiment},
\code{\link[SingleCellExperiment]{SingleCellExperiment}},
\code{\link[SpatialExperiment]{SpatialExperiment}} or
\code{\link[SeuratObject]{SeuratObject}} containing an assay
where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset (row names).}

\item{tumorTissue}{type of tissue for which the signature is developed.}

\item{pvalues}{logical. It allows to compute p-values by permutations.}

\item{nperm}{number of permutations.}

\item{...}{other arguments passed on to the \code{\link[GSVA]{gsva}} function.}
}
\value{
A SummarizedExperiment object in which the passON score
will be added in the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function
 \code{\link[signifinder]{availableSignatures}}.
}
