% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{LICASign}
\alias{LICASign}
\title{Local Immune Cytolytic Activity (LICA) Signature}
\usage{
LICASign(dataset, nametype = "SYMBOL", tumorTissue = "ovary")
}
\arguments{
\item{dataset}{expression values in TPM where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset row names.}

\item{tumorTissue}{tissue of the tumor.}
}
\value{
A SummarizedExperiment object in which the means gene expression based on the mitotix index will be added
in the `colData` section which contains sample meta-data describing the samples.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function \code{\link[signifinder]{availableSignatures}}.
}
