% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{LipidMetSign}
\alias{LipidMetSign}
\title{LIpidSignature}
\usage{
LipidMetSign(dataset, nametype)
}
\arguments{
\item{dataset}{Matrix of expression values where rows correspond to genes and columns correspond to samples.}
}
\description{
Given a dataset, it returns the Lipid score for each sample Mingjun Zheng et all. (2020).
}
