% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ovse}
\alias{ovse}
\title{Ovarian cancer expression data.}
\format{
An object of class \linkS4class{SummarizedExperiment}.
}
\usage{
ovse
}
\description{
A dataset containing gene expression values (in normalized counts, TPM, CPM,
and FPKM) of 40 ovarian cancer patients.
}
\keyword{datasets}
