% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{PiroSign}
\alias{PiroSign}
\title{Piroptosis Signature}
\usage{
PiroSign(dataset, nametype)
}
\arguments{
\item{dataset}{Matrix of expression values where rows correspond to genes and columns correspond to samples. Row names must be Official Symbol.}
}
\description{
Given a dataset, it returns the piroptosis score for each sample, based on Mingjun Zheng et all. (2020).
}
