% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{PlatResSign}
\alias{PlatResSign}
\title{Platinum Resistance Signature}
\usage{
PlatResSign(dataset, nametype, ...)
}
\arguments{
\item{dataset}{Matrix of expression values where rows correspond to genes and columns correspond to samples.}
}
\description{
Given a dataset, it returns the gsva score for each sample from International Cancer Genome Consortium (ICGC).
}
