% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{ridgelineSignPlot}
\alias{ridgelineSignPlot}
\title{Ridgeline Plot}
\usage{
ridgelineSignPlot(data, whichSign = NULL, groupByAnnot = NULL)
}
\arguments{
\item{data}{output from a signature function}

\item{whichSign}{le signature che si vuole far vedere nel ridgeline plot}

\item{groupByAnnot}{un vettore di annotazione dei sample, deve essere categorico}
}
\value{
A ggplot object
}
\description{
Ridgeline Plot
}
