% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signatureTable}
\alias{signatureTable}
\title{Table of all available signatures.}
\format{
An object of class \code{data.frame} with 46 rows and 10 columns.
}
\usage{
signatureTable
}
\description{
A table containing the main information of all the available
signatures in the package.
}
\keyword{datasets}
