% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{SingleSignatureInformation}
\alias{SingleSignatureInformation}
\title{Scatterplot for a single signature}
\usage{
SingleSignatureInformation(data, signatureName, statistics = NULL)
}
\arguments{
\item{data}{either a data frame or a SummarizedExperiment.}

\item{statistics}{It should be "mean", "median" or "quantiles" to be plot on the graph.}

\item{signature}{If data is NULL, signature must be a numeric vector. When data is provided,
signature must be the character string of the column name in data.}
}
\value{
A ggplot object
}
\description{
Given a signatures it returns a scatterplot.
}
