% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subovse}
\alias{subovse}
\title{Example dataset}
\format{
An object of class \code{SummarizedExperiment} with 2911 rows and 40 columns.
}
\usage{
subovse
}
\description{
This dataset is a subset of the ovarian cancer samples collected inside the
cancer genome atlas (TCGA).
}
\keyword{datasets}
