% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{SurvivalPlot}
\alias{SurvivalPlot}
\title{Survival Plot}
\usage{
SurvivalPlot(data, survData, signatureName, cutpoint = "mean")
}
\arguments{
\item{data}{matrice di valori delle signatures}
}
\value{

}
\description{
Survival Plot
}
