% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{survivalSignPlot}
\alias{survivalSignPlot}
\title{Survival Plot}
\usage{
survivalSignPlot(
  data,
  survData,
  signatureName,
  cutpoint = "mean",
  group = NULL,
  groupToUse = NULL
)
}
\arguments{
\item{data}{matrice di valori delle signatures}
}
\value{

}
\description{
Survival Plot
}
