% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{availableSignatures}
\alias{availableSignatures}
\title{Show available signatures}
\usage{
availableSignatures(
  tumorType = NULL,
  tumorTissue = NULL,
  signatureType = NULL,
  inputType = NULL,
  description = TRUE
)
}
\arguments{
\item{tumorTissue}{type of tissue for which the signature is developed}

\item{signatureType}{filter for signature type}

\item{inputType}{type of data you are using: microarray or rnaseq}

\item{description}{whether to show or not the signature description}
}
\value{
a data frame
}
\description{
It shows a table containing all the information of the signatures collected
in the package.
}
