% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{availableSignatures}
\alias{availableSignatures}
\title{Show available signatures}
\usage{
availableSignatures(
  tumor = NULL,
  tissue = NULL,
  topic = NULL,
  requiredInput = NULL,
  description = TRUE
)
}
\arguments{
\item{tumor}{type of tumor for which the signature is developed}

\item{tissue}{type of tissue for which the signature is developed}

\item{topic}{filter for signature topic}

\item{requiredInput}{type of data required: microarray or rnaseq}

\item{description}{whether to show or not the signature description}
}
\value{
a data frame
}
\description{
It shows a table containing all the information of the signatures collected
in the package.
}
\examples{
availableSignatures()

}
