% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{availableSignatures}
\alias{availableSignatures}
\title{Show available signatures}
\usage{
availableSignatures(tumorTissue, signatureType, description)
}
\arguments{
\item{signatureType}{filter for kind of signature}

\item{description}{whether to show or not the signature description}

\item{tissue}{filter for tissue}
}
\value{
a data frame
}
\description{
It shows a table containing all the information of the signatures collected in the package.
}
