% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{CINSign}
\alias{CINSign}
\title{Chromosomal instability Signature}
\usage{
CINSign(
  dataset,
  nametype = "SYMBOL",
  tumorTissue = "breast, lung, brain, lymphatic system",
  typeofCIN = 70
)
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows
correspond to genes and columns correspond to samples.
Alternatively an object of type \linkS4class{SummarizedExperiment},
\code{\link[SingleCellExperiment]{SingleCellExperiment}},
\code{\link[SpatialExperiment]{SpatialExperiment}} or
\code{\link[SeuratObject]{SeuratObject}} containing an assay
where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset (row names).}
}
\value{
A SummarizedExperiment object in which the score will be
added in the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper, to have more
details explore the function \code{\link[signifinder]{availableSignatures}}.
}
