% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{consensusOVSign}
\alias{consensusOVSign}
\title{ConsensusOV Signature}
\usage{
consensusOVSign(
  dataset,
  nametype = "SYMBOL",
  tumorTissue = "ovary",
  method = "consensusOV",
  ...
)
}
\arguments{
\item{dataset}{Expression matrix. A data frame or a matrix where rows correspond to genes and columns correspond to samples.
A SummarizedExperiment which contains an assay represented by a matrix-like object of numeric. The rows typically represent
genomic ranges of interest and the columns represent samples.}

\item{nametype}{gene name ID of your dataset (row names).}

\item{method}{the subtyping method to use. Default is "consensusOV".}
}
\value{
A SummarizedExperiment object in which the COnsensusOV scores will be added
in the `colData` section which contains sample meta-data describing the samples.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function \code{\link[signifinder]{availableSignatures}}.
}
