% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{consensusOVSign}
\alias{consensusOVSign}
\title{ConsensusOV Signature}
\usage{
consensusOVSign(
  dataset,
  nametype = "SYMBOL",
  tumorTissue = "ovary",
  method = "consensusOV",
  ...
)
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows correspond to genes and columns correspond to samples.
Alternatively an object of type \linkS4class{SummarizedExperiment}, \linkS4class{SingleCellExperiment}, \linkS4class{SpatialExperiment} or \linkS4class{Seurat}
containing an assay where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset (row names).}

\item{method}{the subtyping method to use. Default is "consensusOV".}

\item{...}{optional parameters to be passed to \code{\link[consensusOV]{get.subtypes}}.}
}
\value{
A SummarizedExperiment object in which the COnsensusOV scores will be added
in the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function \code{\link[signifinder]{availableSignatures}}.
}
