% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{ConsensusOVSign}
\alias{ConsensusOVSign}
\title{ConsensusOV Signature}
\usage{
ConsensusOVSign(dataset, nametype, method = "consensusOV", ...)
}
\arguments{
\item{dataset}{matrix of expression values where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset row names.}

\item{method}{the subtyping method to use. Default is "consensusOV".}

\item{...}{optional parameters to be passed to the low level function.}
}
\description{
Given a dataset, it returns ovarian cancer subtypes. This signature is based on Chen et. al (2018).
}
