% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{correlationSignPlot}
\alias{correlationSignPlot}
\title{Correlation Plot}
\usage{
correlationSignPlot(
  data,
  whichSign = NULL,
  sampleAnnot = NULL,
  selectByAnnot = NULL
)
}
\arguments{
\item{data}{an object of type \linkS4class{SummarizedExperiment}. Output of
the signatures functions.}

\item{whichSign}{character vector saying the signatures to plot. If not
specified, all the signatures inside data will be plotted. Other signatures
not computed with signifinder can be added in the vector if they are also
included in che colData section of data.}

\item{sampleAnnot}{character vector containing samples' annotations.}

\item{selectByAnnot}{character string saying the subgroup from `sampleAnnot`
used to compute the correlation plot.}
}
\value{
An object of class "openair".
}
\description{
Given multiple signatures, the function plots signatures correlations.
}
\examples{
data(ovse)
correlationSignPlot(data = ovse)

}
