% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{correlationSignPlot}
\alias{correlationSignPlot}
\title{Correlation Plot}
\usage{
correlationSignPlot(
  data,
  whichSign = NULL,
  sampleAnnot = NULL,
  selectByAnnot = NULL
)
}
\arguments{
\item{data}{output from a signature function}

\item{whichSign}{le signature che si vuole far vedere nel correlation plot}

\item{sampleAnnot}{un vettore di annotazione dei sample, deve essere categorico}

\item{selectByAnnot}{a group from sampleAnnot to use to construct the correlation plot}
}
\value{
A correlation ellipse graph
}
\description{
Correlation Plot
}
