% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{correlationSignPlot}
\alias{correlationSignPlot}
\title{Correlation Plot}
\usage{
correlationSignPlot(
  data,
  signatureName = NULL,
  group = NULL,
  groupToUse = NULL
)
}
\arguments{
\item{data}{dataframe con le signature sulle colonne oppure oggetto con i colData}

\item{signatureName}{..}

\item{group}{..}

\item{groupToUse}{..}
}
\value{
A correlation ellipse graph
}
\description{
Correlation Plot
}
