% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{correlationSignPlot}
\alias{correlationSignPlot}
\title{Correlation Plot}
\usage{
correlationSignPlot(
  data,
  whichSign = NULL,
  sampleAnnot = NULL,
  selectByAnnot = NULL
)
}
\arguments{
\item{data}{An object of type \linkS4class{SummarizedExperiment} coming from
signatures functions.}

\item{whichSign}{The signatures to show in the plot.}

\item{sampleAnnot}{A categorical variable containing samples annotations.}

\item{selectByAnnot}{A subgroup from `sampleAnnot` to use to construct the
correlation plot}
}
\value{
A correlation ellipse graph.
}
\description{
Given a matrix of multiple signatures, the function gives the opportunity to
observe signatures correlates between each other.
}
\examples{
OVse <- ferroptosisSign(dataset = subovse)
OVse <- pyroptosisSign(dataset = OVse)
correlationSignPlot(data = OVse)

}
