% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{EMTSign}
\alias{EMTSign}
\title{Endothelial-Mesenchymal Transition Signature}
\usage{
EMTSign(dataset, nametype, ...)
}
\arguments{
\item{dataset}{a matrix of expression values where rows correspond to genes and columns correspond to samples. Row names must be Official Symbol.}

\item{nametype}{gene name ID of your dataset row names.}

\item{...}{other arguments passed on to the GSVA function.}
}
\description{
Given a dataset, it returns the Endothelial score and the Mesenchymal score for each sample, based on QH Miow at all. (2015).
}
