% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{EMTSign}
\alias{EMTSign}
\title{Epithelial-Mesenchymal Transition Signature}
\usage{
EMTSign(
  dataset,
  nametype = "SYMBOL",
  inputType = "microarray",
  author = "Miow",
  pvalues = FALSE,
  nperm = 100,
  ...
)
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows
correspond to genes and columns correspond to samples.
Alternatively an object of type \linkS4class{SummarizedExperiment},
\code{\link[SingleCellExperiment]{SingleCellExperiment}},
\code{\link[SpatialExperiment]{SpatialExperiment}} or
\code{\link[SeuratObject]{SeuratObject}} containing an assay
where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset (row names).}

\item{inputType}{type of data you are using: microarray or rnaseq.}

\item{author}{first author of the specific signature publication.}

\item{pvalues}{logical. It allows to compute p-values by permutations.}

\item{nperm}{number of permutations.}

\item{...}{other arguments passed on to the \code{\link[GSVA]{gsva}} function.}
}
\value{
A SummarizedExperiment object in which the Epithelial and Mesenchymal
scores are added in the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function
 \code{\link[signifinder]{availableSignatures}}.
}
