% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{EMTSign}
\alias{EMTSign}
\title{Epithelial-Mesenchymal Transition Signature}
\usage{
EMTSign(
  dataset,
  nametype = "SYMBOL",
  inputType = "microarray",
  author = "Miow",
  pvalues = FALSE,
  nperm = 100,
  ...
)
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows
correspond to genes and columns correspond to samples.
Alternatively, an object of type \linkS4class{SummarizedExperiment},
\code{\link[SingleCellExperiment]{SingleCellExperiment}} or
\code{\link[SpatialExperiment]{SpatialExperiment}}.}

\item{nametype}{character string saying the type of gene name ID (row names
in dataset). Either one of "SYMBOL", "ENTREZID" or "ENSEMBL".}

\item{inputType}{character string saying the type of data you are using.
Either one of "microarray" or "rnaseq".}

\item{author}{character string saying the first author of the signature
publication. Check it in \code{\link[signifinder]{availableSignatures}}.}

\item{pvalues}{logical. If TRUE it computes p-values of GSVA by permutations.}

\item{nperm}{number of permutations.}

\item{...}{other arguments passed on to the \code{\link[GSVA]{gsva}}
function.}
}
\value{
If dataset is a \linkS4class{SummarizedExperiment} object, then
scores are added in the \code{\link[SummarizedExperiment]{colData}} section.
If dataset is a data frame or a matrix, then a
\linkS4class{SummarizedExperiment} object is created in which scores are
added in the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function
 \code{\link[signifinder]{availableSignatures}}.
}
\examples{
EMTSign(dataset = ovse, inputType = "rnaseq")

}
