% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{EMTSign}
\alias{EMTSign}
\title{Endothelial-Mesenchymal Transition Signature}
\usage{
EMTSign(
  dataset,
  nametype = "SYMBOL",
  tumorTissue = "ovary",
  pvalues = FALSE,
  nperm = 100,
  ...
)
}
\arguments{
\item{dataset}{Expression matrix. A data frame or a matrix where rows correspond to genes and columns correspond to samples.
A SummarizedExperiment which contains an assay represented by a matrix-like object of numeric. The rows typically represent
genomic ranges of interest and the columns represent samples.}

\item{nametype}{gene name ID of your dataset (row names).}

\item{pvalues}{logical. It allows to compute p-values by permutations.}

\item{nperm}{number of permutations.}

\item{...}{other arguments passed on to the GSVA function.}
}
\value{
A SummarizedExperiment object in which the results of the Endothelial score and Mesenchymal score will be added
in the `colData` section which contains sample meta-data describing the samples.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function \code{\link[signifinder]{availableSignatures}}.
}
