% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{EMTSign}
\alias{EMTSign}
\title{Endothelial-Mesenchymal Transition Signature}
\usage{
EMTSign(dataset, nametype = "SYMBOL", pvalues = FALSE, nperm = 100, ...)
}
\arguments{
\item{dataset}{Expression matrix. A data.frame or a matrix where rows correspond to genes and columns correspond to samples.
A SummarizedExperiment which contains an assay represented by a matrix-like object of numeric. The rows typically represent
genomic ranges of interest and the columns represent samples.}

\item{nametype}{gene name ID of your dataset (row names).}

\item{pvalues}{logical. It allows to compute p-values by permutations.}

\item{nperm}{number of permutations.}

\item{...}{other arguments passed on to the GSVA function.}
}
\value{
A SummarizedExperiment object in which the results of the Endothelial score and Mesenchymal score will be added
in the `colData` section which contains sample meta-data describing the samples.
}
\description{
Given a dataset, EMTSign returns the Endothelial score and the Mesenchymal score for
each sample, based on the work of QH Miow at al. (2015).
}
\references{
Miow, Q., Tan, T., Ye, J. et al. Epithelial–mesenchymal status renders differential responses to cisplatin in ovarian cancer.
Oncogene 34, 1899–1907 (2015). \url{https://doi.org/10.1038/onc.2014.136}
}
