% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{ferroptosisSign}
\alias{ferroptosisSign}
\title{FerroptosisSignature}
\usage{
ferroptosisSign(dataset, nametype = "SYMBOL")
}
\arguments{
\item{dataset}{expression values where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset row names.}
}
\description{
Given a dataset, it returns the Ferroptosis score for each sample Ying Ye et al. (2021).
}
