% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{FerrSign}
\alias{FerrSign}
\title{FerroptosisSignature}
\usage{
FerrSign(dataset, nametype)
}
\arguments{
\item{dataset}{matrix of expression values where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of rownames of dataset.}
}
\description{
Given a dataset, it returns the Ferroptosis score for each sample Ying Ye et all. (2021).
}
