% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{geneHeatmapSignPlot}
\alias{geneHeatmapSignPlot}
\title{Gene's Signatures' Heatmap}
\usage{
geneHeatmapSignPlot(
  data,
  whichSign,
  logCount = FALSE,
  splitBySign = FALSE,
  sampleAnnot = NULL,
  splitBySampleAnnot = FALSE,
  ...
)
}
\arguments{
\item{data}{An object of type \linkS4class{SummarizedExperiment} coming from
signatures functions.}

\item{whichSign}{The signatures to show in the plot. This will automatically
select corresponding genes to show in the plot.}

\item{logCount}{Logical. Give the opportunity to compute logarithm for the
expression values in the dataset.}

\item{splitBySign}{Logical. Give you the possibility to split heatmap's row
based on the signature's origin of the genes.}

\item{sampleAnnot}{A vector containing samples annotations.}

\item{splitBySampleAnnot}{A categorical variable. Give you the possibility
to split columns based on samples' annotations.}

\item{...}{Other parameters specific of the function
\code{\link[ComplexHeatmap]{Heatmap}}.}
}
\value{
A \code{\link[ComplexHeatmap]{Heatmap-class}} object.
}
\description{
Given one or multiple signatures, the function gives the opportunity to
observe the trend of the signature's scores based on the expression of the
genes included in each of them.
}
