% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{heatmapSignPlot}
\alias{heatmapSignPlot}
\title{Heatmap globale o con una signature che guida il clustering}
\usage{
heatmapSignPlot(data, signatureName = NULL)
}
\arguments{
\item{data}{dataframe con le signature sulle colonne oppure oggetto con i colData}

\item{signatureName}{una signature (o più) che guidi il clustering}
}
\value{
A ComplexHeatmap object
}
\description{
prende in input una matrice di valori che ha le signatures sulle righe e i pazienti sulle colonne.
}
