% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{heatmapSignPlot}
\alias{heatmapSignPlot}
\title{Heatmap globale o con una signature che guida il clustering}
\usage{
heatmapSignPlot(
  data,
  whichSign = NULL,
  clusterBySign = NULL,
  sampleAnnot = NULL,
  splitByAnnot = FALSE
)
}
\arguments{
\item{data}{output from a signature function}

\item{whichSign}{le signature che si vuole far vedere nell'heatmap}

\item{clusterBySign}{una signature (o più) che guidi il clustering delle colonne dell'heatmap}

\item{sampleAnnot}{un vettore di annotazione dei sample}

\item{splitByAnnot}{se splittare o no le colonne in base all'annotazione dei sample, in
questo caso sampleAnnot deve essere una variabile categorica}
}
\value{
A ComplexHeatmap object
}
\description{
prende in input una matrice di valori che ha le signatures sulle righe e i pazienti sulle colonne.
}
