% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{hypoxiaSign}
\alias{hypoxiaSign}
\title{Hypoxia Signature}
\usage{
hypoxiaSign(dataset, nametype = "SYMBOL", inputType = "microarray")
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows
correspond to genes and columns correspond to samples.
Alternatively an object of type \linkS4class{SummarizedExperiment},
\code{\link[SingleCellExperiment]{SingleCellExperiment}},
\code{\link[SpatialExperiment]{SpatialExperiment}} or
\code{\link[SeuratObject]{SeuratObject}} containing an assay
where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset (row names).}

\item{inputType}{type of data you are using: microarray or rnaseq.}
}
\value{
A SummarizedExperiment object in which the Hypoxia scores is
added in the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function
 \code{\link[signifinder]{availableSignatures}}.
}
\examples{
hypoxiaSign(dataset = ovse, inputType = "rnaseq")

}
