% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{immunoScoreSign}
\alias{immunoScoreSign}
\title{Immunogenic Signature}
\usage{
immunoScoreSign(dataset, nametype = "SYMBOL")
}
\arguments{
\item{dataset}{expression values where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset row names.}
}
\description{
Given a dataset, it returns the ImmunoScore for each sample. This signature is
based on Dapeng Hao et. al (2018).
}
