% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{immunoScoreSign}
\alias{immunoScoreSign}
\title{Immunogenic Signature}
\usage{
immunoScoreSign(
  dataset,
  nametype = "SYMBOL",
  tumorTissue = "ovary",
  author = "Hao"
)
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows correspond to genes and columns correspond to samples.
Alternatively an object of type \linkS4class{SummarizedExperiment}, \linkS4class{SingleCellExperiment}, \linkS4class{SpatialExperiment} or \linkS4class{Seurat}
containing an assay where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset (row names).}
}
\value{
A SummarizedExperiment object in which the Immunogenic scores will be added
in the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function \code{\link[signifinder]{availableSignatures}}.
}
