% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{IPSOVSign}
\alias{IPSOVSign}
\title{IPSOV Signature}
\usage{
IPSOVSign(
  dataset,
  nametype = "SYMBOL",
  inputType = "microarray",
  pvalues = FALSE,
  nperm = 100,
  ...
)
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows
correspond to genes and columns correspond to samples.
Alternatively, an object of type \linkS4class{SummarizedExperiment},
\code{\link[SingleCellExperiment]{SingleCellExperiment}} or
\code{\link[SpatialExperiment]{SpatialExperiment}}.}

\item{nametype}{character string saying the type of gene name ID (row names
in dataset). Either one of "SYMBOL", "ENTREZID" or "ENSEMBL".}

\item{inputType}{character string saying the type of data you are using.
Either one of "microarray" or "rnaseq".}
}
\value{
A SummarizedExperiment object in which scores are added in
the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function
 \code{\link[signifinder]{availableSignatures}}.
}
\examples{
IPSOVSign(dataset = ovse)

}
