% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{IPSSign}
\alias{IPSSign}
\title{ImmunoPhenoScore Signature}
\usage{
IPSSign(dataset, nametype, makeplot = "no", plotpath = NULL)
}
\arguments{
\item{dataset}{TPM expression values where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset row names.}

\item{makeplot}{whether to make or not IPS plots. These will be one for each sample. Default is "no".}

\item{plotpath}{directory path to save plots.}
}
\description{
Given a dataset, it returns IPS for each sample. This signature is based on .. et. al (..).
}
