% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{IPSSign}
\alias{IPSSign}
\title{ImmunoPhenoScore Signature}
\usage{
IPSSign(dataset, nametype = "SYMBOL", hgReference = "hg38")
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows
correspond to genes and columns correspond to samples.
Alternatively, an object of type \linkS4class{SummarizedExperiment},
\code{\link[SingleCellExperiment]{SingleCellExperiment}} or
\code{\link[SpatialExperiment]{SpatialExperiment}}.}

\item{nametype}{character string saying the type of gene name ID (row names
in dataset). Either one of "SYMBOL", "ENTREZID" or "ENSEMBL".}

\item{hgReference}{character string saying the human reference genome.
Either one of "hg19" or "hg38".}
}
\value{
If dataset is a \linkS4class{SummarizedExperiment} object, then
scores are added in the \code{\link[SummarizedExperiment]{colData}} section.
If dataset is a data frame or a matrix, then a
\linkS4class{SummarizedExperiment} object is created in which scores are
added in the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function
 \code{\link[signifinder]{availableSignatures}}.
}
\examples{
data(ovse)
IPSSign(dataset = ovse)

}
