% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{ISCSign}
\alias{ISCSign}
\title{Adult Intestinal Stem Cell Signature}
\usage{
ISCSign(dataset, nametype = "SYMBOL", inputType = "microarray")
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows
correspond to genes and columns correspond to samples.
Alternatively, an object of type \linkS4class{SummarizedExperiment},
\code{\link[SingleCellExperiment]{SingleCellExperiment}},
\code{\link[SpatialExperiment]{SpatialExperiment}} or
\code{\link[SeuratObject]{SeuratObject}}.}

\item{nametype}{character string saying the type of gene name ID (row names
in dataset). Either one of "SYMBOL", "ENTREZID" or "ENSEMBL".}

\item{inputType}{character string saying the type of data you are using.
Either one of "microarray" or "rnaseq".}
}
\value{
A SummarizedExperiment object in which scores are added in
the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function
 \code{\link[signifinder]{availableSignatures}}.
}
\examples{
ISCSign(dataset = ovse, inputType = "rnaseq")

}
