% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{LipidMetSign}
\alias{LipidMetSign}
\title{LipidSignature}
\usage{
LipidMetSign(dataset, nametype = "SYMBOL")
}
\arguments{
\item{dataset}{expression values where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset row names.}
}
\description{
Given a dataset, it returns the Lipid score for each sample Mingjun Zheng et al. (2020).
}
