% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{matrisomeSign}
\alias{matrisomeSign}
\title{Core Matrisome Gene signature}
\usage{
matrisomeSign(dataset, nametype = "SYMBOL")
}
\arguments{
\item{dataset}{expression values where rows correspond to genes and columns correspond to samples.
Row names must be Official Symbol.}

\item{nametype}{gene name ID of your dataset row names.}
}
\description{
Given a dataset, it returns the median genes expression based on Yuzhalin et all. (2018).
}
