% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{matrisomeSign}
\alias{matrisomeSign}
\title{Core Matrisome Gene signature}
\usage{
matrisomeSign(dataset, nametype = "SYMBOL", tumorTissue = "pan-tissue")
}
\arguments{
\item{dataset}{Expression matrix in which row names must be Official Symbol. A data frame or a matrix where rows correspond to genes and columns correspond to samples.
A SummarizedExperiment which contains an assay represented by a matrix-like object of numeric. The rows typically represent
genomic ranges of interest and the columns represent samples.}

\item{nametype}{gene name ID of your dataset (row names).}
}
\value{
A SummarizedExperiment object in which the median gene expression based on the core matrisome signature will be added
in the `colData` section which contains sample meta-data describing the samples.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function \code{\link[signifinder]{availableSignatures}}.
}
