% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{matrisomeSign}
\alias{matrisomeSign}
\title{Core Matrisome Gene signature}
\usage{
matrisomeSign(dataset, nametype = "SYMBOL")
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows
correspond to genes and columns correspond to samples.
Alternatively, an object of type \linkS4class{SummarizedExperiment},
\code{\link[SingleCellExperiment]{SingleCellExperiment}},
\code{\link[SpatialExperiment]{SpatialExperiment}} or
\code{\link[SeuratObject]{SeuratObject}}.}

\item{nametype}{character string saying the type of gene name ID (row names
in dataset). Either one of "SYMBOL", "ENTREZID" or "ENSEMBL".}
}
\value{
A SummarizedExperiment object in which the score will be
added in the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function
 \code{\link[signifinder]{availableSignatures}}.
}
\examples{
matrisomeSign(dataset = ovse)

}
