% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{MetabolicSign}
\alias{MetabolicSign}
\title{Metabolic Signature}
\usage{
MetabolicSign(DEdata, nametype, nsamples)
}
\arguments{
\item{nsamples}{Number of samples in the dataset.}

\item{dataset}{Matrix of differential expression genes where rows correspond to genes, first column correspond to Log2FoldChange and second column to its adjusted pvalue.}
}
\description{
Given a dataset, it returns a matrix with pathways score and a correspondent pvalue calculated with Bootstrapping. This signature is based on Rosario et. all (2018).
}
