% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{metabolicSign}
\alias{metabolicSign}
\title{Metabolic Signature}
\usage{
metabolicSign(DEdata, nametype = "SYMBOL", nsamples)
}
\arguments{
\item{DEdata}{matrix of differentially expressed genes where rows correspond to genes, first column to Log2FoldChange and second column to its adjusted pvalue.}

\item{nametype}{gene name ID of your DEdata row names.}

\item{nsamples}{number of samples in the DEdata.}
}
\description{
Given a list of DEG, it returns a matrix with pathways score and a correspondent pvalue calculated with Bootstrapping.
This signature is based on Rosario et. al (2018).
}
