% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{metabolicSign}
\alias{metabolicSign}
\title{Metabolic Signature}
\usage{
metabolicSign(
  DEdata,
  nametype = "SYMBOL",
  tumorTissue = "pan-tissue",
  nsamples
)
}
\arguments{
\item{DEdata}{A matrix of differentially expressed genes where rows correspond to genes,
the first column to Log2FoldChange and second column to its adjusted p-value,
or a SummarizedExperiment which contains an assay represented by a matrix-like object of differentially expressed genes
where rows correspond to genes, the first column to Log2FoldChange and second column to its adjusted p-value.}

\item{nametype}{gene name ID of your DEdata (row names).}

\item{nsamples}{number of samples in the DEdata.}
}
\value{
A data frame with a Metabolic score for each pathway and the respective p-values.
}
\description{
This signature is computed accordingly to the reference paper,
to have more details explore the function \code{\link[signifinder]{availableSignatures}}.
}
