% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{MitoticIndexSign}
\alias{MitoticIndexSign}
\title{Mitotic Index}
\usage{
MitoticIndexSign(dataset, nametype = "SYMBOL")
}
\arguments{
\item{dataset}{expression values where rows correspond to genes and columns correspond to samples. Row names must be Official Symbol.}

\item{nametype}{gene name ID of your dataset row names.}
}
\description{
Given a dataset, it returns the means genes expression based on Yang et all. (2016).
}
