% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UtilityFunctions.R
\name{multipleSign}
\alias{multipleSign}
\title{Multiple Signature Computation}
\usage{
multipleSign(
  dataset,
  nametype = "SYMBOL",
  inputType = "rnaseq",
  tumorTissue = NULL,
  signatureType = NULL,
  ...
)
}
\arguments{
\item{dataset}{Expression values. A data frame or a matrix where rows
correspond to genes and columns correspond to samples.
Alternatively an object of type \linkS4class{SummarizedExperiment},
\code{\link[SingleCellExperiment]{SingleCellExperiment}},
\code{\link[SpatialExperiment]{SpatialExperiment}} or
\code{\link[SeuratObject]{SeuratObject}} containing an assay
where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset (row names).}

\item{inputType}{type of data you are using: microarray or rnaseq.}

\item{tumorTissue}{compute signatures only from a specific tissue (this
can also be pan-tissue).}

\item{signatureType}{compute signatures only from a specific cancer topic.}

\item{...}{other arguments passed on to the signature functions.}
}
\value{
A SummarizedExperiment object in which the signatures' scores
are added in the \code{\link[SummarizedExperiment]{colData}} section.
}
\description{
This function computes all the signatures for a specific
'inputType' and for a specific 'tumorTissue' and/or 'signatureType'.
}
