% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{oneSignPlot}
\alias{oneSignPlot}
\title{Scatterplot for a single signature}
\usage{
oneSignPlot(data, whichSign, statistics = NULL)
}
\arguments{
\item{data}{An object of type \linkS4class{SummarizedExperiment} coming from
signatures functions.}

\item{whichSign}{The signature function to plot.}

\item{statistics}{The statistics to be plot in the graph. It should be a
pre-defined character between "mean", "median" or "quantiles".}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
Given signatures' scores it returns a scatterplot of patients scores and a
barplot of the density distributions of patients scores.
}
\examples{
OVse <- ferroptosisSign(dataset = sub_OVse)
oneSignPlot(data = OVse, whichSign = "Ferroptosis_Ye")

}
