% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{oneSignPlot}
\alias{oneSignPlot}
\title{Scatterplot for a single signature}
\usage{
oneSignPlot(data, whichSign, statistics = NULL)
}
\arguments{
\item{data}{output from a signature function}

\item{whichSign}{the signature to plot. It has to be only one}

\item{statistics}{It should be "mean", "median" or "quantiles" to be plot in the graph.}
}
\value{
A ggplot object
}
\description{
Given a signatures it returns a scatterplot.
}
