% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{piroptosisSign}
\alias{piroptosisSign}
\title{Piroptosis Signature}
\usage{
piroptosisSign(dataset, nametype = "SYMBOL")
}
\arguments{
\item{dataset}{expression values where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset row names.}
}
\description{
Given a dataset, it returns the piroptosis score for each sample, based on Mingjun Zheng et al. (2020).
}
