% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{PlatResSign}
\alias{PlatResSign}
\title{Platinum Resistance Signature}
\usage{
PlatResSign(dataset, nametype = "SYMBOL", pvalues = FALSE, nperm = 100, ...)
}
\arguments{
\item{dataset}{expression values where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset row names.}

\item{pvalues}{whether to compute pvalues by permutations.}

\item{nperm}{number of permutations.}

\item{...}{other arguments passed on to the GSVA function.}
}
\description{
Given a dataset, it returns the gsva score for each sample from International Cancer Genome Consortium (ICGC).
}
