% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{prognosticSign}
\alias{prognosticSign}
\title{Prognostic Signature}
\usage{
prognosticSign(dataset, nametype = "SYMBOL", age, stage)
}
\arguments{
\item{dataset}{expression values where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset row names.}

\item{age}{vector of patient's age.}

\item{stage}{vector of patient's tumor stage (FIGO).}
}
\description{
Given a dataset, it returns the Quantile assignation for each sample from J. Millstein et. al (2020).
}
