% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SignatureFunction.R
\name{PrognosticSign}
\alias{PrognosticSign}
\title{Prognostic Signature}
\usage{
PrognosticSign(dataset, nametype, age, stage)
}
\arguments{
\item{dataset}{matrix of expression values where rows correspond to genes and columns correspond to samples.}

\item{nametype}{gene name ID of your dataset row names.}

\item{age}{vector of patient's age.}

\item{stage}{vector of patient's tumor stage (FIGO).}
}
\description{
Given a dataset, it returns the Quantile assignation for each sample from J. Millstein et. all (2020).
}
