% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{ridgelineSignPlot}
\alias{ridgelineSignPlot}
\title{Ridgeline Plot}
\usage{
ridgelineSignPlot(data, signatureName = NULL, group = NULL)
}
\arguments{
\item{data}{matrice di valori delle signatures}
}
\value{

}
\description{
Ridgeline Plot
}
