% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{ridgelineSignPlot}
\alias{ridgelineSignPlot}
\title{Ridgeline Plot}
\usage{
ridgelineSignPlot(
  data,
  whichSign = NULL,
  groupByAnnot = NULL,
  selectByAnnot = NULL,
  ...
)
}
\arguments{
\item{data}{An object of type \linkS4class{SummarizedExperiment} coming from
signatures functions.}

\item{whichSign}{The signatures to test in the ridgeline plot.}

\item{groupByAnnot}{A categorical variable containing samples annotations.}

\item{selectByAnnot}{A character indicating the group/s defined in
`groupByAnnot` to show in the plot.}
}
\value{
A \code{\link[ggplot2]{ggplot}} object.
}
\description{
The function calculates densities from the point data mapped onto the x axis,
arranging multiple density plots of the Signatures' scores.
}
\examples{
OVse <- ferroptosisSign(dataset = sub_OVse)
ridgelineSignPlot(data = OVse)

}
