% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifinder-package.R
\docType{package}
\name{signifinder-package}
\alias{signifinder}
\alias{signifinder-package}
\title{signifinder: Implementations of transcriptional cancer signatures}
\description{
signifinder is an R package for computing and exploring a compendium of tumor signatures. It allows computing signatures scores providing the only gene expression values and returns a single-sample score. Currently, signifinder contains 46 distinct signatures collected from the literature.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/CaluraLab/signifinder}
}

}
\author{
\strong{Maintainer}: Stefania Pirrotta \email{stefania.pirrotta@phd.unipd.it}

Authors:
\itemize{
  \item Enrica Calura \email{enrica.calura@unipd.it}
}

}
\keyword{internal}
