% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifinder-package.R
\docType{package}
\name{signifinder-package}
\alias{signifinder}
\alias{signifinder-package}
\title{signifinder: Implementations of transcriptional cancer signatures}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

signifinder is an R package for computing and exploring a compendium of tumor signatures. It allows to compute a variety of signatures, based on gene expression values, and return single-sample scores. Currently, signifinder contains 46 distinct signatures collected from the literature, relating to multiple tumors and multiple cancer processes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/CaluraLab/signifinder}
  \item Report bugs at \url{https://github.com/CaluraLab/signifinder/issues}
}

}
\author{
\strong{Maintainer}: Stefania Pirrotta \email{stefania.pirrotta@phd.unipd.it} (\href{https://orcid.org/0009-0004-0030-217X}{ORCID})

Authors:
\itemize{
  \item Enrica Calura \email{enrica.calura@unipd.it} (\href{https://orcid.org/0000-0001-8463-2432}{ORCID})
}

}
\keyword{internal}
