% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{survivalSignPlot}
\alias{survivalSignPlot}
\title{Survival Plot}
\usage{
survivalSignPlot(
  data,
  survData,
  whichSign,
  cutpoint = "mean",
  sampleAnnot = NULL,
  selectByAnnot = NULL
)
}
\arguments{
\item{data}{output from a signature function}

\item{survData}{dataframe che deve avere due colonne, la prima con i dati di survival (tempo) e la seconda con lo
status, inoltre le righe devono essere nominate con i nomi dei sample, prenderlo dalla documentazione della KM}

\item{whichSign}{la signatura di cui si vuol testare la survival}

\item{cutpoint}{documentazione originale KM}

\item{sampleAnnot}{deve essere un vettore nominato con i nomi dei sample, come le righe, deve essere categorico}

\item{selectByAnnot}{a group from sampleAnnot to use to compute the survival}
}
\value{
al momento ritorna sia il plot che delle statistiche sui dati
}
\description{
Survival Plot
}
