% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{survivalSignPlot}
\alias{survivalSignPlot}
\title{Survival Plot}
\usage{
survivalSignPlot(
  data,
  survData,
  whichSign,
  cutpoint = "mean",
  sampleAnnot = NULL,
  selectByAnnot = NULL
)
}
\arguments{
\item{data}{An object of type \linkS4class{SummarizedExperiment} coming from
signatures functions.}

\item{survData}{A dataframe where each row represents a samples (equal to
the names in the expression matrix) and two columns: the first holding
survival data of time, indicating the follow up time of the samples; the
second holding data of the survival status, an indicator normally 0=alive and
1=dead. For interval censored data, the status indicator is 0=right censored,
1=event at time, 2=left censored, 3=interval censored.}

\item{whichSign}{The signatures to test in the plot.}

\item{cutpoint}{A character between "median", "mean" and "optimal" or a
numeric value, which divide samples between high values and low values. The
function computes the value with the method indicated or employs the values
directly supplied by the user, and based on that number it divides samples in
higher and lower to compare them as the two groups of the survival plot.}

\item{sampleAnnot}{A categorical variable containing samples annotations
named with samples names equal to the row names used in `survData`.}

\item{selectByAnnot}{A group from `sampleAnnot` used to compute the
survival analysis.}
}
\value{
A Survival plot and the statistics computed on data.
}
\description{
It creates survival curves from either a formula (e.g. the
Kaplan-Meier), a previously fitted Cox model, or a previously fitted
accelerated failure time model.
}
\examples{
mysurvData <- cbind(sub_OVse$os, sub_OVse$status)
rownames(mysurvData) <- rownames(colData(sub_OVse))
OVse <- ferroptosisSign(dataset = sub_OVse)
survivalSignPlot(data = OVse,
                 survData = mysurvData,
                 whichSign = "Ferroptosis_Ye")

}
